refactor(agents): co-locate PXI tool definitions behind a sealed registry kernel
Collapse the ~1190-line toolRegistry.ts into a thin aggregator backed by a
small dispatch kernel and per-kind factory helpers:

- registry/defineTool.ts seals each tool's input type inside a closure, so the
  registry holds a homogeneous AgentToolDefinition[] with no `unknown` casts and
  a compiler-checked parser-to-handler link at each definition site.
- registry/defineClientActionTool.ts collapses the duplicated "look up client
  action, bail if unmounted, invoke, map result" boilerplate shared by the
  contextual client-executed tools, with requireSession/buildContext/emitSuccess
  knobs for the approval tools.
- registry/dispatch.ts is the tool-agnostic kernel: server-env guard, name
  lookup, capability gate, then delegation to the tool's dispatch.

Each tool now defines itself in its own module (xAgentTool); toolRegistry.ts
just lists them in an explicit ordered array. Move the previously inlined
set_spans_filter, set_time_range, and render_generative_ui parsers/types into
tools/spansFilter, tools/timeRange, and tools/renderGenerativeUI, add
BASH_TOOL_NAME and ASK_USER_TOOL_NAME constants, and repoint all import sites.

Behavior is unchanged; the existing 21-case registry suite and 241 agent tests
pass. Updates the extending-tool-registry skill doc to the new workflow.
