+++ app/src/agent/tools/bash/constants.ts
@@ -1,0 +1,2 @@
+/** Server-advertised name for the browser-sandboxed bash tool. */
+export const BASH_TOOL_NAME = "bash";
