+++ app/src/agent/tools/codeEvaluatorDraft/index.ts
@@ -1,3 +1,6 @@
+// Code-evaluator form co-pilot tools: read/test inspect the open form, edit
+// proposes a diff the user accepts or rejects (the propose → accept/reject lifecycle).
+export * from "./agentTools";
 export * from "./clientActions";
 export * from "./constants";
 export * from "./draftOperations";
