+++ app/src/agent/tools/elicit/constants.ts
@@ -1,0 +1,2 @@
+/** Server-advertised name for the user-elicitation (ask_user) tool. */
+export const ASK_USER_TOOL_NAME = "ask_user";
