+++ app/src/agent/tools/elicit/index.ts
@@ -1,4 +1,6 @@
 // --- Elicit tool ---
+export { ASK_USER_TOOL_NAME } from "./constants";
+export { askUserAgentTool } from "./askUserAgentTool";
 export { parseElicitToolInput } from "./elicitToolSchema";
 export {
   elicitationOptionSchema,
