+++ app/src/agent/tools/playgroundOutput/index.ts
@@ -1,5 +1,6 @@
 export * from "./clientActions";
 export * from "./constants";
+export * from "./readPlaygroundOutputAgentTool";
 export * from "./parsers";
 export * from "./schemas";
 export * from "./types";
