+++ app/src/agent/tools/playgroundPromptTools/index.ts
@@ -1,3 +1,4 @@
+export * from "./agentTools";
 export * from "./clientActions";
 export * from "./constants";
 export * from "./diffSummary";
