+++ app/src/agent/tools/playgroundRun/index.ts
@@ -1,5 +1,6 @@
 export * from "./clientActions";
 export * from "./constants";
+export * from "./runPlaygroundAgentTool";
 export * from "./parsers";
 export * from "./schemas";
 export * from "./types";
