+++ app/src/agent/tools/playgroundSavePrompt/index.ts
@@ -1,5 +1,6 @@
 export * from "./clientActions";
 export * from "./constants";
+export * from "./savePromptAgentTool";
 export * from "./parsers";
 export * from "./pendingSavePrompt";
 export * from "./savePlaygroundPrompt";
