+++ app/src/agent/tools/playgroundVariableValues/index.ts
@@ -1,5 +1,6 @@
 export * from "./clientActions";
 export * from "./constants";
+export * from "./setVariableValuesAgentTool";
 export * from "./parsers";
 export * from "./schemas";
 export * from "./types";
