+++ app/src/agent/tools/spansFilter/types.ts
@@ -1,0 +1,4 @@
+export type SetSpansFilterInput = {
+  condition: string;
+  rootSpansOnly: boolean;
+};
