+++ app/src/agent/tools/timeRange/types.ts
@@ -1,0 +1,7 @@
+import type { TimeRangeKey } from "@phoenix/components/datetime/types";
+
+export type SetTimeRangeInput = {
+  timeRangeKey: TimeRangeKey;
+  startTime?: string;
+  endTime?: string;
+};
